Implement branch instrumentation for Truffle.
This change adds optional branch instrumentation where executions
of `if` statement branches in specified methods get instrumented with
writes to a special bit array.
On system shutdown, the contents of this array are dumped to the
standard output, showing source location of the `if` statements,
along with information if the `true`, `false` or both branches executed.

A new phase called `InstrumentBranchesPhase` is added to the Truffle partial evaluator.
The graph builder for the Truffle compiler is also modified so that it instruments
`IfNode` objects with `SourceLocation` context information (which wraps a bytecode position).

The change requires slight changes in `BytecodeParser`:

- an `postProcessIfNode` hook is added, which subclasses can override to process the newly created `IfNode`
- the `createBytecodePosition` is made protected