Debugger execution operations need the relevant context
The execution context needs to be passed explicitly, and is part of the corresponding events. In a multithreaded environment, we cannot rely on the thread-local context, especially since UI threads are not having any such context.

This changes DebugExecutionContext from `private` to package private.

TODO: ExecutionEvent needs work, I did not pass the execution context to it, because it has two constructors for which one seems to have lazy semantics, and it is unclear how that could work.
And, this might be in an area where other branches have potentially relevant changes.

Signed-off-by: Stefan Marr <git@stefan-marr.de>