Require that commit messages start with an upper case character
The CONTRIBUTING.md states that each commit message should start with an upper case character. The current check only requires that the first character is not a lower case character. With this change, each commit message has to start with an upper case character as specified by the CONTRIBUTING.md.