Fix gcc dependency in sulong tests and add assembler
Sulong specifies gcc-4.6 (with fortran) as a dependency. During su-tests-sulong the gcc alias is used resulting in failed tests if a matching library set is not installed. Change it to gcc-4.6 to match dependency. Also this commit takes the opportunity to create a Java interface to the LLVM assembler (llvm-as) in the tools project.