Fix double to unsigned int casts
This change fixes casts from double to unsigned int where the double fractional part exceeds the number that can be represented by a *signed* integer.

In such a case we cannot use the standard Java conversion, but first add Integer.MIN_VALUE and then subtract it again. This prevents that floating point values with a fractional part bigger than Integer.MAX_VALUE get converted to Integer.MAX_VALUE.

We do not have this problem with casts from float to unsigned int since the representable fractional part is not big enough to overflow.