Implement the complex conjugate function as a Sulong library function
The Graal NFI currently does not support the calling convention of returning several values. Thus, this change adds a conjugate C function that is executed by Sulong.

This change does not support the float conjugate function yet since we do not support passing vectors to a native function which would be required for the float conjugate.