Null the frame slots before usages
Initially, the frame slot kinds of a `VirtualFrame` are `Object`. The frame slot kinds are not taken from the frame descriptor, to which we assign the statically known types. Instead, they are set the first time when a value is written to the frame.

When compiling code that potentially has two possible frame slot kinds (e.g., when initializing the value in a loop) the Graal graph has phis that select between the `Object` kind and the kind that is assigned later.

To prevent this, this change initializes the frame slots once at the beginning to a 0/null value, so that the frame slot kinds have the right type.