Add possibility to dump graphs as text files
Using `PrintCanonicalGraphStrings=true` switching the graph dumper from
IGV dumping to dumping to text files to a directory structure similar to
the one diplayed in IGV.
`PrintCanonicalGraphStringFlavor` can be used to switch the way graphs are
rendered:
* `0` uses the same format as getCanonicalGraphString used in unit
tests
* `2` uses a format that should make it easier to compare graphs:
The graph is dumped in CFG order, for each block, one line is printed
per fixed node.

Filenames are conservatively escaped.