Flexible selection of a supported LLVM version
Previously, we downloaded LLVM 3.3 binaries that mx used irrespective of installed binaries.
With this change, we first check for supported LLVM binaries (versions 3.2, 3.3, and 3.4) on the system and only download the binaries if we cannot locate supported ones.
This change does not influence the LLVM binaries selected on the Java side.