Fix the linking of external variables
Previously, Sulong used the Graal NFI to link global `extern` variables.
This does not work when another bitcode file defines a `extern` variable, or when we want to intrinsify such a variable.
With this change, we traverse the bitcode libraries and remember defined variables.
When allocating global variables, we can decide based on the defined variables to use the Graal NFI for variables not in this list.