Remove the GCC and LLVM BC test suites and replace them by launching the LL test suites with -Dsulong.TestBinaryParser=true
The GCC and LLVM .bc test suites were introduced with #444 and #449, since the bitcode parser could not parse and execute
all of the .ll GCC and LLVM test cases. Since the bitcode parser can execute all of the test cases that the textual .ll parser
can execute, the separation of the test suites and suite configuration is no longer necessary, and removed with this change.