Add a new mx command to compile testsuites
We plan to add a new test infrastructure that compiles all sources of a suite using mx commands rather than having the logic on the Java side.
The ll files will be picked up by a JUnit test, which runs them with Sulong.
This commit adds basic compilation features to mx and uses them to compile the SulongTestSuite.