Unify the creation of @truffle, @llvm, and other parse-time intrinsics in the factory facade
Previously, there were two different methods for creating parse-time substitutions of @truffle and @llvm functions in the factory facade.
This restricts the extensibility of intrinsic functions, since we would have to create (or remove) additional methods for intrinsics with a new (or deprecated) prefix.
Thus, this change unifies the two methods and also allows the substitution of other functions.