[fix] killCFG: only verify new unused nodes under a special flag
If there were unused nodes in the intial graph it's non-trivial to make
sure there will be no new unused nodes.
Also it shouldn't be a big problem to create new unused nodes.
The flag is still usefull has it might indicate problems in killCFG.
The "unsafe node" verification remains on by default as it should hold
and indicates real problem that can lead to later crashes.