Handle inlining during parsing more like inlining after parsing
When a method has multiple returns, inlinig during parsing created a merge frame state with a bci of one of the return bytecodes. Now the handling is the same as in the inlining phase: the merge frame state that does not include the callee (just the caller frame state after the call).