Fix the inconsistencies in the NodeEventListener.
In the previous design, it was unclear if the event call is supposed to be overridden
by the clients, or the event-specific methods, such as inputChanged,
are supposed to be overridden by the clients.

In the new design, graph-internal modifications always call the event method,
which dispatches to client-overridable methods.
There is an additional method called changed, which is a catch-all.