[GR-7612] Keep state on OSR loop exits
This makes sure we don't deopt to the state before the loop when we
reach the end of the loop. This is particularly important for counted
loop where the profile indicates that the exit is unreached since some
optimization might remove the framestate from the `LoopBegin` of
side-effect-free counted loops.