TRegex: bailout on unsupported lookbehind assertions
Currently, the feature that was tested for was not in exact correspondence
to the actual limits of TRegex can do. Furthermore, before the feature was
checked, the parser was already doing some preprocessing which assumed that
the lookbehind assertions were simple enough for TRegex. This led to broken
assertions in tests which throw complex lookbehind assertions at TRegex.