Fix invalid stamp assignment for AMD64FloatConvertNode.
The problem occurs in the conversion of Math.abs. Stamp folding assigns an i32 stamp restricted
to the integer range to the AbsNode, which is propagated to the FloatConvert node that does the casting.
This is fine until lowering happens, where the FloatConvert gets replaced with a AMD64FloatConvert node.
This node corresponds to the x64 CVTTSS2SI instruction that sometimes returns a special value 0x80000000,
so another snippet that checks the return value, and modifies it, is emitted.
However, this special value is not accounted for in the stamp of the lowered AMD64FloatConvert node,
whose mask remains 0x7fffffff, so canonicalization eventually removes the snippet code.

The reason why this was probably not noticed before is that this pattern occurs rarely.
Usually, the stamp from the AbsNode is either specific enough to fold the node,
or the stamp is completely general, so canonicalization cannot eliminate the snippet.