Add thread-safe handle factory.
JNI fieldID and methodID do not prevent the class (including fields and
method) from being collected.
Espresso uses an efficient, weak collection of handles for methods and
fields.
The handle collection is synchronized for modification but not for
queries, which is the most common operation.
It's not wasteful, will return the same handle for objects already in
the collection.