Minor fixes.
Espresso inherits some fixed properties from the host VM.
This is terribly wrong workaround; most of this properties could be inferred.
e.g. HotSpot derives several properties from the location of libjvm.
But first we have settle on a minimal set properties that need to be provided.
On SVM the location of libmokapot (libvm surrogate) is not enough, since
we cannot derive the sun.boot.class.path, java.home from it.