Canonicalize IfNode to ConditionalNode for variable assignment.
Brief: We want to use "csel" to implement the conditional value selecting operation
for more conditions like Math.min/max, instead of using "cmp + b.cond".

The trueSuccessor and falseSuccessor blocks of IfNode must meet the following conditions:
1. The nodes in block only contain a start and an end one except for the constant and parameter nodes.
2. The merge node only contains a phi member.

All the checking jobs must begin after the FixReadPhase, so that the nodes in blocks will not be changed.

Change-Id: I4aa0ae791e05f34e6af3dd1da035e42a97815b9b