Counted loop exit deopts should not deopt to exit target
This causes deopt loops or failed OSR.

Such deopt branches cna just be duplicated: remove the loop exit and use
a framestate-less BeginNode.
To keep the benefits of recognizing such loops as counted, compute the
"real" exits of loops and use that when detecting counted loops.

Simplify exits computation after FSA (keep old method as an assert for
now).