Use add(extended) to generate array index offset for AArch64.
We want to find a better way to implement array offset calculation.
The old pattern for an int array member loading is:
       and      x4, x3, #0xffffffff
       lsl      x4, x4, #2
       add      x4, x4, #0x10
       ldr      x0, [x2, x4]

And the optimized pattern is:
       add      x1, x2, w3, uxtw #2
       ldr      x0, [x1, #16]

Here is a jmh unit test for int array member loading:
    private final int func(int[] arr) {
        int ret = 0;
        for (int i=0; i < field0; i++) {
            ret += arr[i];
        }
        return ret;
    }
And the jmh performance results:
     without this patch   with this patch   units
         1173501.595        948290.420      us/op

Here are the additional issues fixed related to this patch:
1. Fix an issue in MatchProcessor, so that we can process platform specific node
as a matchable node.

2. Close this optimization for substratevm because the derived reference is not
supported yet in SVM.

mx bootstrap passes.
mx gate -o -t 'CTW:hosted' passes.

Change-Id: Ia34f2bf044ca6863be075cf000b29524ca44c42d