Implement NFAState#equals and NFAState#hashCode using state IDs
In the current implementation, there is at most one NFAState instance per state
ID. Therefore, NFAState object identity is equivalent to NFAState ID equality.
The purpose of this commit is to make NFAState#equals explicit. However, it
could just as well be implemented using object identity, as was the case before.