Use object identity when comparing AST nodes in ASTTransition
When testing AST nodes for equality, we usually want the semantics of object
identity. (Two occurrences of some snippet should be seen as two different pieces
of code. Implementing this without depending on object identity would mean
traversing all of the surrounding context to verify that we are looking at the
same snippet.) Since we want object identity, I have switched from the
equals method to the == operator. This is just refactoring, since the AST nodes
do not override Object#equals.