Trigger killCFG on the unused successors of a newly simplified TypeSwitchNode only *after* the new TypeSwitchNode gets created.
The problem we were encountering was that the new TypeSwitchNode
receives a list of successors, some of which were still reachable,
but getting killed due to other transitive simplifications of the graph.
Concretely, the case was that one of the successors was a LoopExitNode,
that was getting removed due to the loop becoming degenerate during the
deleteBranch call.