Infer stamp for the dividend of integer division and remainder.
This patch creates more opportunities to the following optimizations inside
a counted loop when "d" is a power-of-2 constant:

    "x / d"  ->  "x >> log2(d)"
    "x % d"  ->  "x & (d - 1)"

The above optmization can be applied only if the dividend("x") is known to be
positive. Currently the graal compiler supports this canonicalization. But the
type of the dividend is unrestricted, which means it is not bounded. So most
of the integer divisions and remainders still cannot be canonicalized although
the dividend is positive. So if the compiler can compute the boundary of the
type of the dividend, it can be known as positive, negative or others.

This patch will compute a bounded stamp for the loop variable of a counted loop.
If the variable is used as the dividend of the division or remainder inside of
the loop, the bounded stamp will be used as the real stamp of the dividend.

Directly changing the stamp of the loop variable may generate some issues, for
example it may make a loop endless, or it may influence other loop optimizations.
So this patch just adds a new stamp named "loopStamp" for the counted loop which
represents the stamp of the loop variable only when it is used inside of the loop.
And as I mentioned above, the "loopStamp" will be treated as the new stamp for the
dividend. This operation is applied before the first CanonicalizerPhase, so that
the division and remainder can be optimized as early as possible.

Change-Id: Ia87221897ad67a0ee08bdf4562f357943612426e