AArch64: Better instruction sequence for constant move
Code generation of AArch64 constant move is optimized in this patch.
Original implementation uses MOVZ with MOVK(s) to build all non-zero
16-bit chunks in 32- or 64-bit constants. E.g. Below code is generated
for long integer constant -200000L.

  mov     x1, #0xf2c0
  movk    x1, #0xfffc, lsl #16
  movk    x1, #0xffff, lsl #32
  movk    x1, #0xffff, lsl #48

After this patch, we generate a variable length of instruction sequence
containing MOVZ, MOVN with MOVK(s), based on the value of the constant.
Generated assembly in above example is optimized to below.

  movn    x1, #0xd3f
  movk    x1, #0xfffc, lsl #16

We also renamed AArch64MacroAssemblerTest to AArch64AddressModeTest and
annotateImm to needsImmAnnotation in this patch.