Fix condition for memory bulk zeroing
In the snippet code for allocating new array/object, method fillMemory()
calls ZeroMemoryNode.zero() without checking the value to fill is zero
or not. And, to be more accurate, the value of "size - offset", which is
the actual number of bytes to be zeroed, should be compared against the
value of "MinimalBulkZeroingSize".

Change-Id: Id7b4fc120a81fd50dce3a0abb7b8393efc6dac3a