AArch64 - Read return address correctly.
Since the FrameContext uses `stp fp, lr, ...` to store the
`fp` and `lr` as the first two words in the frame,
the layout is actually:

--------- frame start ---------
lr (word)
fp (word)
--------- rest of frame -------
...

Therefore, when walking the stack, to read the return value
of a given frame with an effective frame-start address, the
immediately first word is the `lr`. Reading the second results
in reading the `fp` which ultimately results in a bus error
on AArch64.