Avoid NPE when parsing static initializers.
Due to the fact an unresolved type may (previous to this commit)
throw a parse exception when attempting to perform automatic unsafe
substitutions, this can result in a null slot being left in the
universe of types array.

Which then ends up being an unresolve type error of type 'null'
subsequently when attempting to resolve null as a type.

Allow the graph-builder configuration to take an optional value
for `allowIncompleteClasspath` which overrides NativeImageOptions
value of the same name. If not provided, use the NativeImageOptions
value (or its default).

Precedence is now:

config.allowIncompleteClasspath (optional) -> NativeImageOptions.AllowIncompleteClasspath (explicit) -> default(false)

Currently only UnsafeAutomaticSubstitutionsProcessor explicitly sets
a value (of `true`) onto the config.