Add `InnocuousThread` to the warning whitelist.
Since it's a JDK class that we can't fix, and probably
can't convince the JDK team to use a field instead of
a local (since a local is all that is required), might as
well whitelist it from the unsafe automatic substitutions
warnings.

Meanwhile, let's spell "suppress" correctly while we're there.