Use memory access stamp of Read/WriteNode to do address lowering.
This patch fixes issue:https://github.com/oracle/graal/issues/1408.

Graal AddressLoweringPhase lowers the general address node to platform
specific address node, like AArch64AddressNode, and together with some
optimizations. It should use the right memory access type to set the
suitable address mode and scaleFactor. But currently it uses the node type.

An AArch64ReadNode is a node merges ZeroExtended/SignedExtended with
ReadNode. If the type of the ReadNode is 32bits. After extending, the
type of the AArch64ReadNode is 64bits, but its access type is still 32bits.
So if the compiler is lowering an AArch64ReadNode and using its node type,
the result address will be wrong.

This patch modifies the stamp of the address to the access stamp.

Change-Id: I0a805134bbff22f9df5626ccdeda95f41ed721e6