Fix HotSpot assertion failure in AArch64 compiler test suite
This fixes issue #1439

For Graal CodeGenTest cases, assembly code generated by assembler or
macro assembler is installed into HotSpot as an nmethod. But AArch64
requires the instruction at nmethod verifed entry to be a jump or nop
for method invalidation in sweeper thread. For this constraint, we
always generate a nop at the beginning in CodeGenTest cases and add an
assertion in AArch64HotSpotBackend.

Change-Id: Ic15f1d502087875c56213fccb1db6fa8c140c4c4