switch from getStackKind to stamp and meta access
We used getStackKind to determine the size of a value (in bytes). As it
turns out, values that are smaller than 4 bytes are padded to be 4 bytes
on the stack. For our application it is not acceptable to treat values
this way since we need to determine properties such as adjacency
accurately.

This commit switches to a different method, which uses the javaType
method on a Stamp and the context's meta access instance. This seems to
give correct results.

The error manifested itself in the BytecodeParser and other parts of the
runtime being unable to load methods during the intrinsics replacement
phase. NoSuchMethodError ended up occurring because methods are looked
up in a hash map that uses the hashCode of two strings.
StringLatin1.hashCode was not compiled properly with vectorization and
exposed that the implementation was not capable of handling types that
are less than 4 bytes large (1 byte in the case of
StringLatin1.hashCode) correctly.