fix loop duplication invalidation bug
LoopTransformations duplicates loop fragments when inserting pre- and
post-loops. One would think that a method that duplicates a fragment
would not modify the original fragment, but the method present here does
just that. Rather than modifying the logic to do with duplication, we
have chosen to instead invalidate the loop's LoopFragmentInside in order
to ensure that both fragments are consistent with each other.

The processPreLoopPhis method processes the LoopFragmentInside and
traverses all of the nodes that are part of this fragment. It asserts
that there is a valid duplicate for each node that is present in the
inside fragment. However, this is not always the case as the duplication
removes some nodes from the LoopFragmentWhole rather than duplicating
them. Since these nodes aren't duplicated, the assertion fails.