Fix ArrayIndexOutOfBoundsException in NewFrameNode
Length of internal array of ArrayList storing slot information
in FrameDescriptor was no longer a good source to determine length,
since it can hold FrameSlots which have larger index than the number
of stored FrameSlots.

Therefore the size of FrameDescriptor is used which only increases with
slot addition therefore may never be smaller than index of a slot.
However, since it is read asynchronously we have to defensively check
that we have actual size to store all the information by slot indexes.