Enable Math.fma intrinsics for AArch64
Math.fma returns the exact product of the first two arguments summed
with the third argument and then rounded once. This behaviour has no
difference with that of the AArch64 FMADD instruction. So this patch
enables Math.fma intrinsics if VM option "UseFMA" is set and the JDK
version is greater than 8.

Change-Id: Iab08fc35b94f4f2dd19277a040c5e92fc01be5a5