Lower thread priority when entering a polyglot thread
Thread priority is lowered when a polyglot thread is entered by an
engine that has the deprioritize flag set. The flag is supported by the
engine builder.

Java thread APIs are used to change the thread priority, vetted by
native code to ensure that the changes are actually reflected by the
scheduling priority of the native threads.