Enable AArch64 travis-ci tests.
Currently the travis-ci of the graal community does not support
tests on AArch64. This could not block patches that might break
AArch64. Fortunately the travis-ci has supported the multi-cpu
architecture tests recently. So it's time to enable the AArch64
tests for graal.

To enable the AArch64 tests and make them run smoothly, this patch
gives the following modifications:
  - Add compiler tests for AArch64.
  - Set the JAVA_HOME if it is not set.
  - Adjust the java version checking rule to make it also work for
    AdoptOpenJDK, which is the openjdk11 installed on travis Arm64.

Change-Id: Ia5cf7f4ea885c62a630917afd1d2cab74c3bbc19