Fixed AArch64 relocation mistakes in SVM.
Currently, in SVM relocation is only enabled for adrp/ldr,adrp/add, and
sequences of moves. However, this limitation was previously not always
observed in SVM code.

Also fixed an error in the backend where it was loading an address into
a register, instead of loading the value at the address

Change-Id: I5b4999e73cbd926c4e877f5f43803a08ed7be5b3