Native Image: Add arch dependent optimization phases.
Currently some architecture specific optimizations of GraalVM are lost in
substratevm, like the AArch64ReadReplacementPhase for AArch64. In GraalVM,
they are appended either to the suites or to the LIR suites list.

This patch fixes it by using the platform specific suites creator to create
the Mid-End optimization suites and LIR suites.

Change-Id: Ia9b4f26c456311513bba40841221905e11ce3beb