Update supported max java version to 15.
The aot and graalunit jtreg tests fail when run with graal after
the java version of the jdk/jdk master being updated to 15. This
is caused by the class file version checking in Graal. It should
be inside the defined java version range. Currently the max java
version is defined to 14, which should be updated to 15.

Change-Id: I4431ceda3ea2f62efdc5ebee3c8ba13c319b4a4d