Added support for actions on transition to Java.
Added a thread local variable to indicate that an action should be
performed before a thread begins to execute Java code. This variable is
checked and the action is triggered as needed by each thread when its
state returns to the STATUS_IN_JAVA state.

This functionality is needed on AArch64 to allow the processor pipeline
to be flushed by all threads after either new code is installed or prior
code is patched.

Change-Id: Ic7207d2252d4ee0db85e5933481cd0e8bf28d7ae