Explain why implicitExceptionBytes may be null
Because not all versions of the JDK have the field aot_metadata._nul_chk_table_begin, you do not always
want to generate this field depending on the target JDK's version. This difference is taken care of by
HotSpotGraalServices which has different implementations based on the target JDK's version. The
HotSpotGraalServices.getImplicitExceptionBytes returns null in case the field does not exist.

Additionally, harmonize how exceptionBytes is generated with the other variables.