Put exception initialization behind boundaries.
In the Espresso codebase the exception klasses are always constant.
This is a very nice property since it allows to PE through exception
allocation and initialization.
In practice, this is not enough; even if the exception class is
explicit, the meta accessor may not be PE constant.
The dynamism is sometimes non-deliberate, e.g. when obtaining the
meta accessor from the receiver... where the receiver klass is not
PE constant.