Update Arm copyright regex.
Currently the copyright regex for Arm part assumes that it has only one
year recorded. Considering that we might update the AArch64 files and
together with appending the latest modification year after the created
year, we should make the copyright regex be matched for other valid pattern.
Or it will report warning like "copyright that does not match the regex".
This issue has happened when we execute "mx updategraalinopenjdk" and the
copyright of one AArch64 file contains "Copyright (c) 2019, 2020, Arm Limited.".

Change-Id: I0a8dc283886150c07286dc4f5efe1e2f2c0f5656