Apply trapping nullcheck to uncompressed object for compressed pointer.
Trapping nullcheck might generate two uncompress instructions for
the same compressed oop on AArch64. One is inserted by the backend
when it emits nullcheck. If the object is a compressed pointer, it
is uncompressed before the nullcheck is emitted. And another one is
generated by the uncompression node used for memory access. These
two instructions are duplicated with each other.

The generated codes on AArch64 like:

  ldr   w0, [x0,#112]
  lsl   x2, x0, #3      ; uncompressing (first)
  ldr   xzr, [x2]       ; implicit exception: deoptimizes
  ......                ; fixed operations
  lsl   x0, x0, #3      ; uncompressing (second)
  str   w1, [x0,#12]

A simple way to avoid this is to creat a new uncompression node for
the nullcheck, and let the value numbering remove the duplicated one
if possible. Since the address lowering of AMD64 can handle the
uncompressing computation for address, the created uncompression node
is wrapped to an address node and the nullcheck is finally applied on
the address.

With the modification, the codes above could be optimized to:

  ldr   w0, [x0,#112]
  lsl   x0, x0, #3      ; uncompressing
  ldr   xzr, [x0]       ; implicit exception: deoptimizes
  ......                ; fixed operations
  str   w1, [x0,#12]