AArch64: jaotc generates incorrect code for compressed OOPs with non-zero heap base
On AArch64 if the VM jaotc is running in uses compressed oops with a
zero base it will not emit instructions to add/subtract the heap base
when compressing/decompressing oops. This causes a crash if the AOT'd
shared library is loaded into a VM with non-zero heap base.

Fix for JDK-8244164.

Change-Id: I9983f6ac81b5ef4a91ce595f61455b90ba1546c3