Make sharable nodes be consumable by matched rules
This patch fixes issue: https://github.com/oracle/graal/issues/2216

Unused constant load won't be deleted after match rules. As a result,
it might generate redundant "mov/orr" instructions if the constants
cannot be inlined to codes.

For example, the original codes:

  lsr    w0, w0, #6
  orr    w1, wzr, #0xfffff
  and    w0, w0, w1

will be optimized to the following pattern by adding match rules:

  orr    w1, wzr, #0xfffff
  ubfx   w0, w0, #6, #20

However, the constant "0xfffff" is not consumed by the matched rules
since ConstantNode is sharable.

To fix it, this patch adds the modifications to the matching codes:
  1) Adds the "consumable" flag for matching. It is used to decide
whether a node can be consumed by a matched rule.
  2) Adds the class "SharedMatchValue". It is used as the operand for
a sharable and consumable node, if the node is swallowed by the matched
rules. It doesn't emit any codes.

Change-Id: Ibed66714c12e02f400ab494683dee8826ede01a1