AArch64: Add the full barrier implementation support for CAS.
Graal always uses "LDAXR+STLXR" for Unsafe atomic operations. However
this is not sufficient on AArch64 for the VM internal atomics that need
the full barrier semantics (See [1]).

This patch adds an alternative implementation for "CAS", which has the
full barrier semantics. The final "CAS" pattern is decided by an flag
"useBarriers". Since we prefer to use the optimized "LDAXR+STLXR" pattern
, the flag is false by default. It can be set to true by checking the
memory "LocationIdentity" for the VM internal "CAS" operation if it needs
the full barrier semantic.

[1] https://patchwork.kernel.org/patch/3575821/

Change-Id: Ifa57c9733b1dd1a2e2208e885a11932475a4e3b0