Add tests, covering `fitsIn{floating point type}`.
1. Checking if an integer value fits into a floating point type
Tests verify that whenever a floating point type has enough precision to hold
an integer value, the conversion succeeds. In addition, if the floating point type
is not precise enough, but the integer ends with enough zeroes, the conversion
succeeds. However, if the floating point type is not precise enough and the value
ends with 1, the conversion should fail. Add the corresponding checks to the
interop implementation as well.

2. Checking if a float value fits into double