Fix SimpleLanguage readMember implementation
As far as I can tell, readMember implementations must assume that they
may be called with invalid keys (i. e., the caller is not responsible
for calling isMemberReadable first), and should themselves throw an
UnknownIdentifierException if necessary rather than return null.

The specific pattern used here – return the value if it is not null,
unconditional throw at the end – is based on the readMember
implementations of WasmModule and PythonAbstractObject (though both of
them have more than one potential return before the throw).