UnsafeSubstitutionsTest stores object to int field
Running this test on AArch64 with -XX:-UseCompressedOops crashes the VM
with a SIGBUS. This is caused by a call to unsafePutObjectVolatile with
a non-64-bit-aligned offset (12). The VM should throw an exception here
rather than crash, and that needs to be fixed separately (it's a variant
of JDK-8246051), but we should also fix the test to use the offset of
the correct field.

Change-Id: I67677b207cb5792b0ba758a8d61ae0109e64127b