Add intrinsic for StringUTF16.indexOfCharUnsafe and StringLatin1.indexOf
This patch implements intrinsics for String.indexOf  (only for searching
a char, not substring). The intrinsic uses Graal snippets for its
implementation. The generated code is similar to that of C2 for UTF-16
strings, and the same approach is adapted for Latin1 strings.
  - For the small strings (less than 8-bytes in length), we search char
by char which is similar to the interpreter's approach.
  - For longer strings, we search by chunk of size 8-bytes.
The snippets are replaced with a foreign function call that align with
the implementation of x86. Introducing a function call has a small
overhead for small strings that is then compansated as the length of the
string increases.
Microbenchmarking on a A72 system shows speedup up to 2.1x for long
strings (128 chars). For small strings (6 chars) the speedup is 0.69x.

Change-Id: If778e3fd8c7297a82a60560cfedb751f00d3e23b