Remove alternative constant-store policies.
This simplifies interpreter implementation at the instructions that have
to decode constants.

We still however pay an overhead when accessing multi-byte constants,
because the LEB128 decoding loop is not unrolled.
The problem is that the loop is not counted, because we cannot prove
that there will be no out-of-bounds errors when accessing the
instruction array.
We should therefore manually read 2-3 64-bit values and decode accordingly,
but that can be done later.