Sulong: Support system calls that return pointers
This works by introducing a second specialization type
signature. Since most syscalls return a long and the special case is
returning pointers (e.g. mmap). The executeLong specialization will be
executed first by the Truffle DSL and in case it returns an Object,
the UnexpectedResultException will contain that object as a result,
and the DSL will rewrite the call to executeGeneric from this point
on. If a long is returned (as expected), nothing will change.