Enhance MethodEntries creation and verification
Following Andrew Dinn's recommendation in
https://github.com/oracle/graal/pull/3304#issuecomment-815926794, this
patch creates the necessary `MethodEntry`s for each `ClassEntry` using
the methodInfoProvider, It then sorts the methods list to ensure faster
lookups in the phase of linking `Range`s to `MethodEntry`s.

It also adds assertions to ensure that the methods list is sorted and
that there are no duplicate entries in it.