CI: Test with Quarkus main instead of Quarkus latest release
Quarkus current latest release is 1.13.1.Final which does not support
GraalVM 21.1. As discussed in
https://github.com/oracle/graal/issues/3272 GraalVM 21.1 brings some
changes that break several Quarkus tests. Quarkus is now preparing its
2.x release which is expected to take a bit longer than its usual
releases. This patch makes the CI use Quarkus `main` branch instead of
the latest release, as `main` is meant to work with the upcoming GraalVM
21.1 release.