Use method entries instead of primary ranges for method DIE generation
As of https://github.com/oracle/graal/pull/3304 we associate every range
with the `MethodEntry` of the method it was compiled from, and as a
result we know that a `MethodEntry` exists for each compiled method in
the image. Actually we create `MethodEntry`s even for methods that are
not compiled but are part of `ClassEntry` that describes a Class
reachable from in the image.

For every range we generate a DIE (Debugging Information Entry) which
references another DIE that holds the specification of the method that
this range was generated from. So for every range we need a method DIE
containing information for the corresponding method.

Till now we have been generating these DIEs by traversing the primary
ranges of each ClassEntry. There are, however, subranges that might be
compiled from a method that is not referenced by a primary range, e.g.,
inlined methods.

This patch ensures that every method that is compiled in the native
image (including inlined ones) will get a method DIE.