Update FileEntry of MethodEntry in setInRange
If the MethodEntry was added by traversing the DeclaredMethods of a
Class its fileEntry will point to the original source file, thus it will
be wrong for substituted methods. As a result when setting a MethodEntry
as isInRange we also make sure that its fileEntry reflects the file info
associated with the corresponding Range.