Simplify array writes of reinterpret nodes
A ReinterpretNode does not change the bit pattern of its input value.
For an array we can write any number of bytes at a time without messing
up the LocationIdentity. As a result this change allows a
ReinterpretNode to be folded away when writing to an array location - we
just write the original input value. This is beneficial to SIMD values
to save moving values betwen GPR and vector registers.