Add SIMD implementation of String.indexOf char on AArch64
This patch replaces the chunk-based char search algorithm
with its SIMD equivalent. It uses chunks of size 32 bytes
instead of 4 bytes. The patch repurposed the char search
approach of Arm optimised routines [1].

The benefits of using SIMD implementation should increase
for longer strings. For long strings (~750 Latin1 chars),
microbenchmarking shows an improvement of 8% (on A72) and
33% (on Neoverse N1).

[1] https://github.com/ARM-software/optimized-routines/blob/master/string/aarch64/strchr.S

Change-Id: I83f80ccff473b07bcb839111d0cffdbaae9a9098